requirements.txt
anyio==4.0.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
comm==0.1.4
constantly==15.1.0
contourpy==1.1.1
cryptography==39.0.0
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.3.0
executing==2.0.1
Faker==19.13.0
fastjsonschema==2.18.1
filelock==3.12.4
Flask==2.2.2
fonttools==4.43.1
fqdn==1.5.1
frozendict==2.3.8
functions-framework==3.3.0
google==3.0.0
greenlet==2.0.1
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.26.0
ipython==8.16.1
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
jwt==1.3.1
kafka-python==2.0.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
multitasking==0.0.11
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
overrides==7.4.0
packaging==23.0
pandas==2.1.2
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
Protego==0.3.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pygame==2.3.0
Pygments==2.16.1
pymongo==4.4.0
PyMySQL==1.0.2
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyspark==3.5.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
Scrapy==2.5.0
scrapy-rotating-proxies==0.6.2
seaborn==0.13.0
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spark==0.2.1
SQLAlchemy==1.4.46
stack-data==0.6.3
terminado==0.17.1
tinycss2==1.2.1
tldextract==5.0.1
tornado==6.3.3
traitlets==5.12.0
Twisted==22.10.0
types-python-dateutil==2.8.19.14
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.3
w3lib==2.1.2
watchdog==2.2.1
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.2.2
yfinance==0.2.32
zope.interface==6.1

confluent_kafka==2.4.0
py4j==0.10.9.7
pyspark==3.5.0 
elasticsearch==8.13.0
